fix: currency converter exchange rate type error

+++ src/main/java/main/arbitrage/common/util/currency/CurrencyConverter.java
@@ -7,12 +7,12 @@ import java.math.RoundingMode;
 
 @Component
 public class CurrencyConverter {
-    public BigDecimal krwToUsd(BigDecimal krw, BigDecimal exchangeRate) {
-        return krw.divide(exchangeRate, 4, RoundingMode.HALF_UP);
+    public BigDecimal krwToUsd(BigDecimal krw, double exchangeRate) {
+        return krw.divide(BigDecimal.valueOf(exchangeRate), 4, RoundingMode.HALF_UP);
     }
 
-    public BigDecimal usdToKrw(BigDecimal usd, BigDecimal exchangeRate) {
-        return usd.multiply(exchangeRate).setScale(0, RoundingMode.HALF_UP);
+    public BigDecimal usdToKrw(BigDecimal usd, double exchangeRate) {
+        return usd.multiply(BigDecimal.valueOf(exchangeRate)).setScale(0, RoundingMode.HALF_UP);
     }
 }
